+++ src/main/java/duke/ui/Ui.java
@@ -11,6 +11,10 @@ public class Ui {
 
     private TaskList tasks;
 
+    /**
+     * This method shows welcome message to user.
+     *
+     */
     public void welcome(){
         System.out.println("------------------------------------------------------------");
         String logo = " ____        _        \n"
@@ -25,22 +29,41 @@ public class Ui {
         System.out.println("------------------------------------------------------------");
     }
 
+    /**
+     * This method reads user input.
+     *
+     */
     public String readCommand(){
         Scanner in = new Scanner(System.in);
         String command = in.nextLine().trim();
         return command;
     }
 
+    /**
+     * This method prints exit message for user.
+     *
+     */
     public static void exitCommand() {
         printCommand("Bye. Hope to see you again soon!");
     }
 
+    /**
+     * This method prints custom message for user.
+     *
+     * @param command A String representing custom messages
+     */
     public static void printCommand(String command) {
         System.out.println("------------------------------------------------------------");
         System.out.println("\t" + command);
         System.out.println("------------------------------------------------------------");
     }
 
+
+    /**
+     * This method prints list of tasks.
+     *
+     * @param taskList A TaskList object representing the list of tasks.
+     */
     public static void printList(TaskList taskList) {
         String printStr = "";
         for (int i = 0; i < taskList.getSize(); i++) {
@@ -55,6 +78,12 @@ public class Ui {
         }
     }
 
+    /**
+     * This method prints deleted task and number of tasks left in the task list.
+     *
+     * @param taskList A TaskList object representing the list of tasks.
+     * @param task A Task object representing the deleted task.
+     */
     public static String getDeletedTaskString(TaskList taskList, Task task) {
         String addedTask = "Noted. I've removed this task: " + System.lineSeparator();
         addedTask += "\t\t" + task + System.lineSeparator();
@@ -63,6 +92,12 @@ public class Ui {
     }
 
 
+    /**
+     * This method prints new task added and number of tasks in the task list.
+     *
+     * @param taskList A TaskList object representing the list of tasks.
+     * @param task A Task object representing the new task.
+     */
     public static String getNewTaskString(TaskList taskList, Task task) {
         String addedTask = "Got it. I've added this task: " + System.lineSeparator();
         addedTask += "\t\t" + task + System.lineSeparator();
